+++ benches/internals.rs
@@ -91,6 +91,19 @@ pub fn criterion_benchmark(c: &mut Criterion) {
         })
     });
 
+    c.bench_function("updates into_route_iter", |b| {
+        b.iter(|| {
+            let mut reader = black_box(&updates[..]);
+
+            BgpkitParser::from_reader(&mut reader)
+                .into_route_iter()
+                .take(RECORD_LIMIT)
+                .for_each(|x| {
+                    black_box(x);
+                });
+        })
+    });
+
     c.bench_function("updates into_raw_record_iter", |b| {
         b.iter(|| {
             let mut reader = black_box(&updates[..]);
@@ -136,6 +149,19 @@ pub fn criterion_benchmark(c: &mut Criterion) {
 
             BgpkitParser::from_reader(&mut reader)
                 .into_update_iter()
+                .take(RECORD_LIMIT)
+                .for_each(|x| {
+                    black_box(x);
+                });
+        })
+    });
+
+    c.bench_function("rib into_route_iter", |b| {
+        b.iter(|| {
+            let mut reader = black_box(&rib_dump[..]);
+
+            BgpkitParser::from_reader(&mut reader)
+                .into_route_iter()
                 .take(RECORD_LIMIT)
                 .for_each(|x| {
                     black_box(x);